feat: #99 extension 수정
